requirements.txt
GitPython==3.0.5
PyYAML==5.2
streamlit==0.52.2
